requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.2.0
ecdsa==0.19.1
Flask==3.1.1
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
